[ar71xx] bootdrv: enable support to download caldata
This change enables the support for the download of caldata from
the host board to the target board.  To differentiate between the
actual 2nd stage firmware and the caldata, the host will set the
execution address as zero for caldata (for the 2nd stage firmware,
the host will set a valid execution address).

Change-Id: I15197cb5ad6b2b09b4e2ecfe26afb4528b18bdd3
Signed-off-by: Prabhu Jayakumar <pjayak@codeaurora.org>